flyttet riktekst
